refactoring done
1. Found two methods that both returned -1 in SudokuVerifier Class.
Used Consolidate conditionals technique and extracted a method
testOfLengthAndNumbers which encapsulates both methods that returned
same error code.

2. Used Jdeodorant on SudokuVerifierTest class and it found smells
considering God classes. I used Jdeodorant automatic extraction tool
which  extracted class SudokuVerifierTestRowsAndNumbers. The automatic
refactoring did the extraction because not all the string attributes
inside SudokuVerifierTest were used by all of the methods.

3. Also removed the extra methods from parametrized test class.
