+++ .settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+encoding//tests/SudokuVerifierTestRowsAndNumbers.java=Cp1252
